requirements.txt
langchain-openai
langchain-community
tiktoken
scikit-learn
numpy
pandas
transformers
python-dotenv
flask
logging